+++ src/configuration/models/uniteTheme/uniteThemeConfiguration.ts
@@ -10,6 +10,6 @@ export class UniteThemeConfiguration {
     public themeHeaders: string[];
     public backgroundColor: string;
     public themeColor: string;
-    public appLoaderStyle: string;
-    public appLoader: string;
+    public appLoaderStyle: string[];
+    public appLoader: string[];
 }