+++ src/pipelineSteps/unite/uniteThemeConfigurationJson.ts
@@ -52,13 +52,29 @@ export class UniteThemeConfigurationJson extends PipelineStepBase {
         defaultConfiguration.themeHeaders = [];
         defaultConfiguration.backgroundColor = "#339933";
         defaultConfiguration.themeColor = "#339933";
-        defaultConfiguration.appLoaderStyle = "<style>#app-loader{width:200px;height:200px;position:absolute;top:0;bottom:0;left:0;right:0;margin:auto;}</style>";
-        defaultConfiguration.appLoader = `<svg width="200px" height="200px" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" preserveAspectRatio="xMidYMid">
-            <circle cx="50" cy="50" fill="none" stroke="{THEME_COLOR}" stroke-width="2" r="35"
-                stroke-dasharray="164.93361431346415 56.97787143782138" transform="rotate(330 50 50)">
-                <animateTransform attributeName="transform" type="rotate" calcMode="linear"
-                    values="0 50 50;360 50 50" keyTimes="0;1" dur="1s" begin="0s" repeatCount="indefinite">
-            </animateTransform></circle></svg>`;
+        defaultConfiguration.appLoaderStyle = [
+            "<style>",
+            "#app-loader",
+            "{",
+            "width:200px;",
+            "height:200px;",
+            "position:absolute;",
+            "top:0;",
+            "bottom:0;",
+            "left:0;",
+            "right:0;",
+            "margin:auto;",
+            "}",
+            "</style>"
+        ];
+        defaultConfiguration.appLoader = [
+            `<svg width="200px" height="200px" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" preserveAspectRatio="xMidYMid">`,
+            `<circle cx="50" cy="50" fill="none" stroke="{THEME_COLOR}" stroke-width="2" r="35" stroke-dasharray="164.93361431346415 56.97787143782138" transform="rotate(330 50 50)">`,
+            `<animateTransform attributeName="transform" type="rotate" calcMode="linear" values="0 50 50;360 50 50" keyTimes="0;1" dur="1s" begin="0s" repeatCount="indefinite">`,
+            `</animateTransform>`,
+            `</circle>`,
+            `</svg>`
+        ];
 
         this._configuration = ObjectHelper.merge(defaultConfiguration, this._configuration);