+++ test/unit/src/pipelineSteps/content/htmlTemplate.spec.ts
@@ -87,7 +87,7 @@ describe("HtmlTemplate", () => {
             Chai.expect(loggerInfoSpy.args[0][0]).contains("Writing");
 
             const lines = await fileSystemMock.fileReadLines("./test/unit/temp/www/", "index-no-bundle.html");
-            Chai.expect(lines.length).to.be.equal(18);
+            Chai.expect(lines.length).to.be.equal(19);
         });
 
         it("can write if file has a generated marker", async () => {
@@ -106,7 +106,7 @@ describe("HtmlTemplate", () => {
             Chai.expect(loggerInfoSpy.args[0][0]).contains("Writing");
 
             const lines = await fileSystemMock.fileReadLines("./test/unit/temp/www/", "index-no-bundle.html");
-            Chai.expect(lines.length).to.be.equal(20);
+            Chai.expect(lines.length).to.be.equal(21);
             Chai.expect(lines.findIndex(line => line.indexOf("head1") > 0)).to.be.equal(7);
             Chai.expect(lines.findIndex(line => line.indexOf("body2") > 0)).to.be.equal(13);
         });